Splits prediction error plots from residuals (#1078)
This PR splits prediction error plots from residuals, closes #1031
