+++ docs/api/regressor/peplot.rst
@@ -6,8 +6,8 @@ Prediction Error Plot
 A prediction error plot shows the actual targets from the dataset against the predicted values generated by our model. This allows us to see how much variance is in the model. Data scientists can diagnose regression models using this plot by comparing against the 45 degree line, where the prediction exactly matches the model.
 
 =================   ==============================
- Visualizer           :class:`~yellowbrick.regressor.residuals.PredictionError`
- Quick Method         :func:`~yellowbrick.regressor.residuals.prediction_error`
+ Visualizer           :class:`~yellowbrick.regressor.prediction_error.PredictionError`
+ Quick Method         :func:`~yellowbrick.regressor.prediction_error.prediction_error`
  Models               Regression
  Workflow             Model Evaluation
 =================   ==============================
@@ -16,8 +16,9 @@ A prediction error plot shows the actual targets from the dataset against the pr
     :context: close-figs
     :alt: Prediction Error plot on the Concrete dataset using a linear model
 
-    from sklearn.model_selection import train_test_split
     from sklearn.linear_model import Lasso
+    from sklearn.model_selection import train_test_split
+
     from yellowbrick.datasets import load_concrete
     from yellowbrick.regressor import PredictionError
 
@@ -43,8 +44,9 @@ The same functionality can be achieved with the associated quick method ``predic
     :context: close-figs
     :alt: Prediction Error plot on the Concrete dataset using a linear model
 
-    from sklearn.model_selection import train_test_split
     from sklearn.linear_model import Lasso
+    from sklearn.model_selection import train_test_split
+
     from yellowbrick.datasets import load_concrete
     from yellowbrick.regressor import prediction_error
 
@@ -61,7 +63,7 @@ The same functionality can be achieved with the associated quick method ``predic
 API Reference
 -------------
 
-.. automodule:: yellowbrick.regressor.residuals
+.. automodule:: yellowbrick.regressor.prediction_error
     :members: PredictionError, prediction_error
     :undoc-members:
     :show-inheritance:
