+++ tests/baseline_images/test_regressor/test_prediction_error/test_peplot_no_lines.png
