+++ tests/baseline_images/test_regressor/test_prediction_error/test_prediction_error_quick_method.png
